% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{base_model}
\alias{base_model}
\title{Function to create baseline prediction and compute excess}
\usage{
base_model(
  time_case,
  T,
  years,
  morData,
  sexCol,
  ageCol,
  popCol,
  timeCol,
  use.rate = FALSE
)
}
\arguments{
\item{time_case}{string indicating monthly or weekly input}

\item{T}{number of time periods}

\item{years}{a sequence of years to show}

\item{morData}{the long format data frame}

\item{sexCol}{column name for sex variable}

\item{ageCol}{column name for age variable}

\item{popCol}{column name for population variable}

\item{timeCol}{column name for within-year time variable}

\item{use.rate}{predict based on rate}
}
\value{
a list of calculated excess mortality
}
\description{
Function to create baseline prediction and compute excess
}
\examples{
data(SampleInput1)
SampleInput1$sex <- SampleInput1$age <- "All"
out <- base_model(time_case = "Monthly", T = 12, years = c(2015:2021), morData = SampleInput1, 
			  sexCol = "sex", ageCol = "age", popCol = "population", timeCol = "month")
data(SampleInput2)
out <- base_model(time_case = "Monthly", T = 12, years = c(2015:2021), morData = SampleInput2, 
			  sexCol = "sex", ageCol = "age", popCol = "population", timeCol = "month")
data(SampleInput3)
out <- base_model(time_case = "Weekly", T = 53, years = c(2015:2021), morData = SampleInput3, 
			  sexCol = "sex", ageCol = "age", popCol = "population", timeCol = "week")
}
