% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcessMortalityApp.R
\docType{data}
\name{SampleInput1}
\alias{SampleInput1}
\title{Example dataset without sub-population breakdown}
\format{
A data frame with 4 columns
}
\description{
The dataset should contain the following columns: year, month, deaths, population.
}
\examples{

data(SampleInput1)

}
\keyword{datasets}
