% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcessMortalityApp.R
\docType{data}
\name{SampleInput3}
\alias{SampleInput3}
\title{Example weekly dataset with sex and age breakdown}
\format{
A data frame with 6 columns
}
\description{
The dataset should contain the following columns: year, month, deaths, sex, age, population.
}
\examples{

data(SampleInput3)

}
\keyword{datasets}
