% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_model.R
\name{smooth_model}
\alias{smooth_model}
\title{Function to create baseline prediction and compute excess using time series model}
\usage{
smooth_model(
  time_case,
  T,
  years,
  morData,
  sexCol,
  ageCol,
  popCol,
  timeCol,
  use.rate = TRUE,
  knots_per_year = 4,
  return_model = FALSE
)
}
\arguments{
\item{time_case}{string indicating monthly or weekly input}

\item{T}{number of time periods}

\item{years}{a sequence of years to show}

\item{morData}{the long format data frame}

\item{sexCol}{column name for sex variable}

\item{ageCol}{column name for age variable}

\item{popCol}{column name for population variable}

\item{timeCol}{column name for within-year time variable}

\item{use.rate}{predict based on rate}

\item{knots_per_year}{number of AR1 knots per year.}

\item{return_model}{logical indicator to print the model summary}
}
\value{
a list of calculated excess mortality
}
\description{
Function to create baseline prediction and compute excess using time series model
}
\examples{
library(INLA)
data(SampleInput1)
SampleInput1$sex <- SampleInput1$age <- "All" 
out <- smooth_model(time_case = "Monthly", T = 12, years = c(2015:2021), 
			  morData = SampleInput1, 
			  sexCol = "sex", ageCol = "age", 
			  popCol = "population", timeCol = "month")
mortality_plot(model = out, sex = "All", age = "All", timeCol = "month",
				month_or_week = "Monthly", plot_show = "Death Counts")
mortality_plot(model = out, sex = "All", age = "All", timeCol = "month", 
				month_or_week = "Monthly", plot_show = "Excess Death Counts")
 
data(SampleInput2)
out <- smooth_model(time_case = "Monthly", T = 12, years = c(2015:2021), 
			  morData = SampleInput2, 
			  sexCol = "sex", ageCol = "age", 
			  popCol = "population", timeCol = "month")
mortality_plot(model = out, sex = "All", age = "65+", timeCol = "month", 
				month_or_week = "Monthly", plot_show = "Death Counts")
mortality_plot(model = out, sex = "Female", age = "65+", timeCol = "month", 
				month_or_week = "Monthly", plot_show = "Death Counts")
mortality_plot(model = out, sex = "Male", age = "65+",  timeCol = "month",
				month_or_week = "Monthly", plot_show = "Excess Death Counts")

data(SampleInput3)
SampleInput3$sex <- SampleInput3$age <- "All"
out <- smooth_model(time_case = "Weekly", T = 52, years = c(2015:2021), 
			  morData = SampleInput3, 
			  sexCol = "sex", ageCol = "age", 
			  popCol = "population", timeCol = "week")
mortality_plot(model = out, sex = "All", age = "All", timeCol = "week",
				month_or_week = "Weekly", plot_show = "Death Counts")
mortality_plot(model = out, sex = "All", age = "All", timeCol = "week",
				month_or_week = "Weekly", plot_show = "Excess Death Counts")

data(SampleInput3)
out <- smooth_model(time_case = "Weekly", T = 53, years = c(2015:2021), 
			  morData = SampleInput3, 
			  sexCol = "sex", ageCol = "age", 
			  popCol = "population", timeCol = "week", return_model=TRUE)
summary(out$model$all[[1]])
}
