% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Function to create summary tables}
\usage{
summary_table(time_case, T, years, morData)
}
\arguments{
\item{time_case}{string indicating monthly or weekly input}

\item{T}{number of time periods}

\item{years}{a sequence of years to show}

\item{morData}{the long format data frame}
}
\value{
a list of summary tables
}
\description{
Function to create summary tables
}
\examples{
data(SampleInput1)
SampleInput1$sex <- SampleInput1$age <- "All"
SampleInput1$timeCol = SampleInput1$month
SampleInput1$popCol = SampleInput1$population
tab <- summary_table(SampleInput1, time_case = "Monthly", T = 12, years = 2015:2021)
data(SampleInput3)
SampleInput3$timeCol = SampleInput3$week
SampleInput3$popCol = SampleInput3$population
tab <- summary_table(SampleInput3, time_case = "Weekly", T = 53, years = 2015:2021)

}
