% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot.R
\name{compare_plot}
\alias{compare_plot}
\title{Function to plot multiple excess mortality estimates}
\usage{
compare_plot(model, by, month_or_week, plot_show, timeCol = "timeCol")
}
\arguments{
\item{model}{output from base_model}

\item{by}{character variable for "By Sex", "By Age", or "By Sex and Age"}

\item{month_or_week}{character variable for monthly or weekly input}

\item{plot_show}{character variable for which plot to show}

\item{timeCol}{column specifying time variable}
}
\value{
a ggplot object
}
\description{
Function to plot multiple excess mortality estimates
}
\examples{
data(SampleInput3)
out <- base_model(time_case = "Weekly", T = 53, 
			  years = c(2015:2021), morData = SampleInput3, 
			  sexCol = "sex", ageCol = "age", 
			  popCol = "population", timeCol = "week")
compare_plot(model = out, by = "By Sex",timeCol = "week",
			  month_or_week = "Weekly", 
           plot_show = "Death Counts")
compare_plot(model = out, by = "By Age",timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Death Counts")
compare_plot(model = out, by = "By Sex and Age", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Death Counts")
compare_plot(model = out, by = "By Sex", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Excess Death Counts")
compare_plot(model = out, by = "By Age", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Excess Death Counts")
compare_plot(model = out, by = "By Sex and Age", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Excess Death Counts")
compare_plot(model = out, by = "By Sex", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Excess Death Counts (Overlay)")
compare_plot(model = out, by = "By Age", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Excess Death Counts (Overlay)")
compare_plot(model = out, by = "By Sex and Age", timeCol = "week",
           month_or_week = "Weekly", 
           plot_show = "Excess Death Counts (Overlay)")
}
